REFACTOR : 컴포넌트 경로 리팩토링
